requirements.txt
google-cloud-storage >= 1.17.0